+++ src/cpp/session/modules/SessionFindTests.cpp
@@ -32,7 +32,7 @@ using namespace modules::find;
 namespace {
 
 const std::string kLine("RStudio is great");
-const std::string replaceString("awesome");
+const std::string kReplaceString("awesome");
 const size_t matchOn = 11;
 const size_t matchOff = 16;
 
@@ -40,9 +40,11 @@ const std::string kRegexLine("aba OOOkkk okab AAOO aaabbb aa abab");
 const size_t rMatchOn = 4;
 const size_t rMatchOff = 10;
 
-const std::string findRegex("\\([a-z]\\)\\1\\{2\\}\\([a-z]\\)\\2\\{2\\}");
-const std::string replaceRegex("\\1\\2\\1\\2");
+const std::string kFindRegex("\\([a-z]\\)\\1\\{2\\}\\([a-z]\\)\\2\\{2\\}");
+const std::string kReplaceRegex("\\1\\2\\1\\2");
 
+const std::string kGrepPattern("aba \033[01m\033[KOOOkkk\033[m\033[K okab AAOO awesome aa abab");
+const std::string kGitGrepPattern("aba \033[1;31mOOOkkk\033[m okab AAOO awesome aa abab");
 } // anonymous namespace
 
 TEST_CASE("SessionFind")
@@ -54,7 +56,7 @@ TEST_CASE("SessionFind")
 
       Replacer replacer(true);
       replacer.replaceLiteral(matchOn, matchOff,
-         replaceString, &line, &replaceMatchOff);
+         kReplaceString, &line, &replaceMatchOff);
       CHECK(line.compare("RStudio is awesome") == 0);
       CHECK(replaceMatchOff == 18);
    }
@@ -66,7 +68,7 @@ TEST_CASE("SessionFind")
 
       Replacer replacer(false);
       replacer.replaceRegex(rMatchOn, rMatchOff,
-         findRegex, replaceString, &line, &replaceMatchOff);
+         kFindRegex, kReplaceString, &line, &replaceMatchOff);
       CHECK(line.compare("aba OOOkkk okab AAOO awesome aa abab") == 0);
       CHECK(replaceMatchOff == 36);
    }
@@ -77,14 +79,14 @@ TEST_CASE("SessionFind")
       size_t replaceMatchOff;
 
       Replacer replacer(true);
-      replacer.replaceRegex(rMatchOn, rMatchOff, findRegex, replaceRegex, &line,
+      replacer.replaceRegex(rMatchOn, rMatchOff, kFindRegex, kReplaceRegex, &line,
          &replaceMatchOff);
       CHECK(line.compare("aba OkOk okab AAOO aaabbb aa abab") == 0);
       CHECK(replaceMatchOff == 8);
 
       size_t matchOn = 18;
       size_t matchOff = 24;
-      replacer.replaceRegex(matchOn, matchOff, findRegex, replaceRegex, &line,
+      replacer.replaceRegex(matchOn, matchOff, kFindRegex, kReplaceRegex, &line,
          &replaceMatchOff);
       CHECK(line.compare("aba OkOk okab AAOO abab aa abab") == 0);
       CHECK(replaceMatchOff == 22);
@@ -96,7 +98,7 @@ TEST_CASE("SessionFind")
       size_t replaceMatchOff;
 
       Replacer replacer(false);
-      replacer.replaceRegex(rMatchOn, rMatchOff, findRegex, replaceRegex, &line,
+      replacer.replaceRegex(rMatchOn, rMatchOff, kFindRegex, kReplaceRegex, &line,
          &replaceMatchOff);
       CHECK(line.compare("aba OOOkkk okab AAOO abab aa abab") == 0);
       CHECK(replaceMatchOff == 33);
@@ -108,7 +110,7 @@ TEST_CASE("SessionFind")
       size_t replaceMatchOff = 99;
 
       Replacer replacer(true);
-      Error error = replacer.replaceRegex(rMatchOn, rMatchOff, findRegex, replaceRegex,
+      Error error = replacer.replaceRegex(rMatchOn, rMatchOff, kFindRegex, kReplaceRegex,
          &line, &replaceMatchOff); 
       CHECK(line.compare(kLine) == 0);
    }
@@ -152,6 +154,48 @@ TEST_CASE("SessionFind")
       replacer.replaceRegex(on, off, findPattern, replacePattern, &line, &replaceMatchOff);
       CHECK(line.compare("hello world") == 0);
       CHECK(replaceMatchOff == 11);
+   }
+
+   SECTION("Grep get file, line number, and contents")
+   {
+      boost::regex regex = getGrepOutputRegex(/*isGitGrep*/ false);
+      std::string contents(
+         "case.test:2:aba \033[01m\033[KOOOkkk\033[m\033[K okab AAOO awesome aa abab");
+
+      boost::smatch match;
+      CHECK(regex_utils::match(contents, match, regex));
+      CHECK(match[1].str().compare("case.test") == 0);
+      CHECK(match[2].str().compare("2") == 0);
+      CHECK(match[3].str().compare(kGrepPattern) == 0);
+   }
+
+   SECTION("Grep get color encoding regex")
+   {
+      boost::regex regex = getColorEncodingRegex(/*isGitGrep*/ false);
+      boost::cmatch match;
+      CHECK(regex_utils::search(kGrepPattern.c_str(), match, regex));
+      CHECK(match[1].str().compare("01") == 0);
+   }
+
+   SECTION("Git grep get file, line number, and contents")
+   {
+      boost::regex regex = getGrepOutputRegex(/*isGitGrep*/ true);
+      std::string contents(
+   "case.test\033[36m:\033[m2\033[36m:\033[maba \033[1;31mOOOkkk\033[m okab AAOO awesome aa abab");
+
+      boost::smatch match;
+      CHECK(regex_utils::match(contents, match, regex));
+      CHECK(match[1].str().compare("case.test") == 0);
+      CHECK(match[2].str().compare("2") == 0);
+      CHECK(match[3].str().compare(kGitGrepPattern) == 0);
+   }
+
+   SECTION("Git grep get color encoding regex")
+   {
+      boost::regex regex = getColorEncodingRegex(/*isGitGrep*/ true);
+      boost::cmatch match;
+      CHECK(regex_utils::search(kGitGrepPattern.c_str(), match, regex));
+      CHECK(match[2].str().compare("1") == 0);
    }
 }