+++ src/gwt/src/org/rstudio/core/client/widget/ModalDialogBase.java
@@ -1,7 +1,7 @@
 /*
  * ModalDialogBase.java
  *
- * Copyright (C) 2009-19 by RStudio, Inc.
+ * Copyright (C) 2009-20 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -84,7 +84,6 @@ public abstract class ModalDialogBase extends DialogBox
       // a11y
       role_ = role;
       role_.set(getElement());
-      A11y.setARIADialogModal(getElement());
 
       // main panel used to host UI
       mainPanel_ = new VerticalPanel();