+++ src/gwt/src/org/rstudio/studio/client/application/Application.java
@@ -351,7 +351,7 @@ public class Application implements ApplicationEventHandlers
    @Override
    public void onAriaLiveStatus(AriaLiveStatusEvent event)
    {
-      int delayMs = userPrefs_.get().typingStatusDelayMs().getValue();
+      int delayMs = event.getImmediate() ? 0 : userPrefs_.get().typingStatusDelayMs().getValue();
       if (!ModalDialogTracker.dispatchAriaLiveStatus(event.getMessage(), delayMs))
          view_.reportStatus(event.getMessage(), userPrefs_.get().typingStatusDelayMs().getValue());
    }