+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTargetWidget.java
@@ -1,7 +1,7 @@
 /*
  * TextEditingTargetWidget.java
  *
- * Copyright (C) 2009-19 by RStudio, Inc.
+ * Copyright (C) 2009-20 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -189,7 +189,7 @@ public class TextEditingTargetWidget
                      clamped = 0;
                   
                   editorPanel_.setWidgetSize(docOutlineWidget_, clamped);
-                  toggleDocOutlineButton_.setLatched(clamped != 0);
+                  setDocOutlineLatchState(clamped != 0);
                   editor_.onResize();
                }
                
@@ -212,10 +212,12 @@ public class TextEditingTargetWidget
             editorPanel_,
             statusBar_);
       
+      Roles.getTabpanelRole().set(panel_.getElement());
+      setAccessibleName(null);
       adaptToFileType(fileType);
       
       editor.addFocusHandler(event ->
-            toggleDocOutlineButton_.setLatched(docOutlineWidget_.getOffsetWidth() > 0));
+            setDocOutlineLatchState(docOutlineWidget_.getOffsetWidth() > 0));
 
       editor_.setTextInputAriaLabel("Text editor");
 
@@ -243,7 +245,7 @@ public class TextEditingTargetWidget
          double widgetSize = target_.getPreferredOutlineWidgetSize();
          double size = Math.min(editorSize, widgetSize);
          editorPanel_.setWidgetSize(docOutlineWidget_, size);
-         toggleDocOutlineButton_.setLatched(true);
+         setDocOutlineLatchState(true);
       }
    }
    
@@ -504,6 +506,7 @@ public class TextEditingTargetWidget
       toggleDocOutlineButton_ = new LatchingToolbarButton(
             ToolbarButton.NoText,
             ToolbarButton.NoTitle,
+            true, /* textIndicatesState */
             new ImageResource2x(StandardIcons.INSTANCE.outline2x()),
             event -> {
                final double initialSize = editorPanel_.getWidgetSize(docOutlineWidget_);
@@ -523,7 +526,7 @@ public class TextEditingTargetWidget
                   title = title.replace("Hide ", "Show ");
                toggleDocOutlineButton_.setTitle(title);
 
-               toggleDocOutlineButton_.setLatched(destination != 0);
+               setDocOutlineLatchState(destination != 0);
 
                int duration = (userPrefs_.reducedMotion().getValue() ? 0 : 500);
                new Animation()
@@ -562,7 +565,7 @@ public class TextEditingTargetWidget
                   ? title.replace("Show ", "Hide ")
                   : title.replace("Hide ", "Show ");
             toggleDocOutlineButton_.setTitle(title);
-            toggleDocOutlineButton_.setLatched(docOutlineWidget_.getOffsetWidth() > 0);
+            setDocOutlineLatchState(docOutlineWidget_.getOffsetWidth() > 0);
          }
       }.schedule(100);
       
@@ -586,6 +589,12 @@ public class TextEditingTargetWidget
       toolbar.addRightWidget(showWhitespaceCharactersCheckbox_);
       
       return toolbar;
+   }
+   
+   private void setDocOutlineLatchState(boolean latched)
+   {
+      toggleDocOutlineButton_.setLatched(latched);
+      docOutlineWidget_.setAriaVisible(latched);
    }
    
    private ToolbarButton createLatexFormatButton()
@@ -773,7 +782,7 @@ public class TextEditingTargetWidget
       if (!fileType.canShowScopeTree())
       {
          editorPanel_.setWidgetSize(docOutlineWidget_, 0);
-         toggleDocOutlineButton_.setLatched(false);
+         setDocOutlineLatchState(false);
       }
       
       toolbar_.invalidateSeparators();
@@ -812,6 +821,14 @@ public class TextEditingTargetWidget
    public void setNotebookUIVisible(boolean visible)
    {
       runSetupChunkOptionMenu_.setVisible(visible);
+   }
+
+   @Override
+   public void setAccessibleName(String name)
+   {
+      if (StringUtil.isNullOrEmpty(name))
+         name = "Untitled Text editor";
+      Roles.getTabpanelRole().setAriaLabelProperty(panel_.getElement(), name);
    }
 
    public HasValue<Boolean> getSourceOnSave()