+++ src/gwt/src/org/rstudio/studio/client/workbench/views/vcs/dialog/graph/GraphLine.java
@@ -1,7 +1,7 @@
 /*
  * GraphLine.java
  *
- * Copyright (C) 2009-12 by RStudio, Inc.
+ * Copyright (C) 2009-20 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -28,6 +28,7 @@ public class GraphLine
       columns_ = new GraphColumn[vals.length];
       for (int i = 0; i < columns_.length; i++)
          columns_[i] = new GraphColumn(vals[i]);
+      altText_ = "";
    }
 
    public GraphColumn[] getColumns()
@@ -53,6 +54,7 @@ public class GraphLine
    {
       draw(s_canvas, theme);
       return SafeHtmlUtil.createOpenTag("img",
+                                        "alt", altText_,
                                         "class", theme.getImgClassName(),
                                         "src", s_canvas.toDataUrl());
    }
@@ -108,6 +110,7 @@ public class GraphLine
             if (c.nexus)
             {
                nexusColumn = i;
+               altText_ = "commit depth " + nexusColumn;
                ctx.setFillStyle(theme.getColorForId(c.id));
             }
 
@@ -151,6 +154,7 @@ public class GraphLine
    }
 
    private GraphColumn[] columns_;
+   private String altText_;
 
    // Use a static canvas to avoid the overhead of continually recreating them
    private static final Canvas s_canvas = Canvas.createIfSupported();